add some sanity checking to scripts

+++ scripts/reboot.sh
@@ -1,4 +1,10 @@
 #!/bin/bash
 
+if [ "$(id -u)" -ne "0" ]
+then
+	echo "This script must be run as root."
+	exit 1
+fi
+
 sleep 2
 reboot
